libuntl/requirements/dev.txt
Werkzeug==0.11.15
ipython	
django-extensions
flake8
